# swagger-collection.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Apartment Management API v1.0",
    "version": "1.0.0",
    "description": "OpenAPI 3 — import vào Swagger Editor / UI. Mọi JSON body dùng schema có `example` trên từng field. Các UUID trong ví dụ có dạng ngẫu nhiên (giống `Guid.NewGuid()` khi seed) và **không** phải id thật trong DB — sau khi seed, thay bằng giá trị từ API. JWT: login rồi gửi Authorization: Bearer. Seed data range: apartments BLK-0001..BLK-0150, bulk users demoblk00..demoblk39@seed.local, canonical accounts admin1/admin2 and user1..user6 @ apartment.local. UUID examples are placeholders; after seeding, lấy ID thật từ API list/detail."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1.0/Users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Current user profile",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update my profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMyProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Users/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user (Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CurrentUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Soft delete user (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Users/{id}/purge": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
        }
      ],
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Purge user permanently (Admin)",
        "description": "Removes Identity user, refresh tokens, authored feedbacks, and unlinks residents.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Users/{id}/password": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
        }
      ],
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Admin reset user password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetUserPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1.0/Residents/me": {
      "get": {
        "tags": [
          "Residents"
        ],
        "summary": "My resident record",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Residents/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
        }
      ],
      "get": {
        "tags": [
          "Residents"
        ],
        "summary": "Get resident (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Residents"
        ],
        "summary": "Update resident (Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResidentUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Residents"
        ],
        "summary": "Delete resident (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Residents/{id}/restore": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
        }
      ],
      "post": {
        "tags": [
          "Residents"
        ],
        "summary": "Restore resident (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdateMyProfileDto": {
        "type": "object",
        "required": [
          "fullName"
        ],
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Resident Demo 40"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "example": "+84901654321"
          },
          "currentPassword": {
            "type": "string",
            "nullable": true,
            "example": "ExampleFormat!1Abc"
          },
          "newPassword": {
            "type": "string",
            "nullable": true,
            "example": "NewStr0ng!Pass"
          }
        }
      },
      "CurrentUserDto": {
        "type": "object",
        "required": [
          "userId",
          "email",
          "fullName",
          "roles"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "8f3e2c1a-9b5d-4f7e-a162-3c4d5e6f7089"
          },
          "email": {
            "type": "string",
            "example": "demoblk39@seed.local"
          },
          "fullName": {
            "type": "string",
            "example": "Resident Demo 40"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "example": "+84901654321"
          },
          "avatarPath": {
            "type": "string",
            "nullable": true,
            "example": "/uploads/avatars/..."
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Admin"
            ]
          }
        }
      },
      "ResetUserPasswordDto": {
        "type": "object",
        "required": [
          "newPassword"
        ],
        "properties": {
          "newPassword": {
            "type": "string",
            "example": "AdminReset!456"
          }
        }
      },
...
      "ResidentCreateDto": {
        "type": "object",
        "required": [
          "apartmentId",
          "fullName",
          "phoneNumber",
          "isOwner",
          "moveInDate"
        ],
        "properties": {
          "apartmentId": {
            "type": "string",
            "format": "uuid",
            "example": "8f3e2c1a-9b5d-4f7e-a162-3c4d5e6f7089"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "example": "7d2c5e9b-1e4f-4a8c-b243-5f6a7b8c9012"
          },
          "fullName": {
            "type": "string",
            "example": "Guest resident 149"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+84901654321"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "guest149@seed.local"
          },
          "isOwner": {
            "type": "boolean",
            "example": true
          },
          "moveInDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-15T00:00:00.000Z"
          }
        }
      },
      "ResidentUpdateDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ResidentCreateDto"
          },
          {
            "type": "object",
            "properties": {
              "moveOutDate": {
                "type": "string",
                "format": "date-time",
                "nullable": true,
                "example": null
              }
            }
          }
        ]
      },
...
    }
  }
}